Checking for presence of key before adding value
Doing some tidying up and adding code to check for the presence of a key before adding it to the HashMap. At every key, there is an ArrayList which contains an scheduleInformation object containing all the data about that particular schedule. This prevents any values from being overwritten if there are two schedules with the same arrival times in the file stop_times.txt